+++ src/pages/PostEditPage.tsx
@@ -1,0 +1,5 @@
+import { PostEditContent } from "@/features/board/components";
+
+export default function PostEditPage() {
+  return <PostEditContent />;
+}
